Task 7: Bonus (Optional) Global state provided react use
