+++ package.json
@@ -22,6 +22,7 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-router-dom": "^6.16.0",
+		"react-use": "^17.4.0",
 		"sort-by": "^1.2.0"
 	},
 	"devDependencies": {
