+++ src/App.tsx
@@ -4,7 +4,7 @@ const App = () => {
   return (
 		<>
 			<Counter />
-			<Counter />
+			{/* <Counter /> */}
 		</>
 	);
 }
